Mannnnn... i UnoReceiver stod det att den tar emot en 16-bitars array... det ska vara en 8-bitars array... ffs...

+++ RTOS V.6/UnoReceiver/UnoReceiver.ino
@@ -97,7 +97,7 @@ Pick_Up_Status pick_up_status_t = PICK_UP_IDLE;
 TWI_CMD twi_cmd_t = TWI_CMD_NONE;
 
 /* Buffers for receiving and transmitting bytes */
-uint16_t rx_buf[RX_DATA_LENGTH];
+uint8_t rx_buf[RX_DATA_LENGTH];
 uint8_t tx_buf[TX_DATA_LENGTH];
 
 void setup()
